# test/resources/openapi.yaml
openapi: '3.0.0'
info:
  version: 1.0.0
  title: Swagger Petstore
  description: A sample API that uses a petstore as an example to demonstrate features in the OpenAPI 3.0 specification
  termsOfService: http://swagger.io/terms/
  contact:
    name: Swagger API Team
    email: apiteam@swagger.io
    url: http://swagger.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /v1/
  - url: http://{name}.swagger.io:{port}/{version}
    variables:
      name:
        default: petstore
        enum:
          - petstore
          - storeofpets
      port:
        enum:
          - '443'
          - '8443'
        default: '443'
      version:
        default: v1
        enum:
          - v1
          - v2
paths:
  /pets:
    description: endpoints for pets
    summary: endpoints for pets
    get:
      description: |
        Returns all pets from the system that the user has access to
        Nam sed condimentum est. Maecenas tempor sagittis sapien, nec rhoncus sem sagittis sit amet. Aenean at gravida augue, ac iaculis sem. Curabitur odio lorem, ornare eget elementum nec, cursus id lectus. Duis mi turpis, pulvinar ac eros ac, tincidunt varius justo. In hac habitasse platea dictumst. Integer at adipiscing ante, a sagittis ligula. Aenean pharetra tempor ante molestie imperdiet. Vivamus id aliquam diam. Cras quis velit non tortor eleifend sagittis. Praesent at enim pharetra urna volutpat venenatis eget eget mauris. In eleifend fermentum facilisis. Praesent enim enim, gravida ac sodales sed, placerat id erat. Suspendisse lacus dolor, consectetur non augue vel, vehicula interdum libero. Morbi euismod sagittis libero sed lacinia.

        Sed tempus felis lobortis leo pulvinar rutrum. Nam mattis velit nisl, eu condimentum ligula luctus nec. Phasellus semper velit eget aliquet faucibus. In a mattis elit. Phasellus vel urna viverra, condimentum lorem id, rhoncus nibh. Ut pellentesque posuere elementum. Sed a varius odio. Morbi rhoncus ligula libero, vel eleifend nunc tristique vitae. Fusce et sem dui. Aenean nec scelerisque tortor. Fusce malesuada accumsan magna vel tempus. Quisque mollis felis eu dolor tristique, sit amet auctor felis gravida. Sed libero lorem, molestie sed nisl in, accumsan tempor nisi. Fusce sollicitudin massa ut lacinia mattis. Sed vel eleifend lorem. Pellentesque vitae felis pretium, pulvinar elit eu, euismod sapien.
      operationId: findPets
      parameters:
        - name: tags
          in: query
          description: tags to filter by
          required: false
          style: form
          schema:
            type: array
            items:
              type: string
        - name: limit
          in: query
          description: maximum number of results to return
          required: true
          schema:
            type: integer
            format: int32
            minimum: 5
        - name: test
          in: query
          description: maximum number of results to return
          required: true
          schema:
            type: string
            enum:
              - one
              - two
        - name: testJson
          in: query
          description: JSON in query params
          content:
            application/json:
              schema:
                type: object
                properties:
                  foo:
                    type: string
                    enum:
                      - bar
                      - baz
        - name: testArrayNoExplode
          in: query
          description: Array in query param
          schema:
            type: array
            items:
              type: string
              enum:
                - foo
                - bar
                - baz
          style: form
          explode: false
        - name: testArrayNoExplode2
          in: query
          description: Array in query param
          schema:
            type: array
            items:
              type: string
          style: form
          explode: false
        - name: testArrayExplode
          in: query
          description: Array explode in query param
          schema:
            type: array
            items:
              type: string
              enum:
                - foo
                - bar
                - baz
          style: form
      #          explode: true
      responses:
        '200':
          description: pet response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      description: Creates a new pet in the store.  Duplicates are allowed
      operationId: addPet
      requestBody:
        description: Pet to add to the store
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPet'
      responses:
        '200':
          description: pet response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /pets/{id}:
    get:
      description: Returns a user based on a single ID, if the user does not have access to the pet
      operationId: find pet by id
      parameters:
        - $ref: 'xt.openapi.parameters.yaml#/parameters/id'
        # - name: id
        #   in: path
        #   description: ID of pet to fetch
        #   required: true
        #   schema:
        #     type: integer
        #     format: int64
      responses:
        '200':
          description: pet response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      description: deletes a single pet based on the ID supplied
      operationId: deletePet
      parameters:
        - name: id
          in: path
          description: ID of pet to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: pet deleted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # /pets/{id}/form_urlencoded:
  #   post:
  #     requestBody:
  #       content:
  #         application/x-www-form-urlencoded:
  #           schema:
  #             type: object
  #             required:
  #               - form_p1
  #             properties:
  #               form_p1:
  #                 type: string
  #               form_p2:
  #                 type: string
  #     responses:
  #       '200':
  #         description: photo uploaded
  /pets/{id}/photos:
    post:
      description: upload a photo of the pet
      operationId: uploadPetPhoto
      parameters:
        - name: id
          in: path
          description: ID of pet to fetch
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - metadata
              properties:
                file:
                  description: The photo
                  type: string
                  format: binary
                metadata:
                  type: string
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - form_p1
              properties:
                form_p1:
                  type: string
                form_p2:
                  type: string
      responses:
        '200':
          description: photo uploaded

  /pets/{id}/attributes:
    get:
      description: Returns attributes for this pet
      operationId: findAttributes
      parameters:
        - name: id
          in: path
          description: ID of pet to fetch
          required: true
          schema:
            type: integer
            format: int64
        - in: header
          name: X-Attribute-Id
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '200':
          description: pet response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attribute'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /pets/{id}/attributes/{attribute_id}:
    get:
      description: Returns the attribute specified by attribute_id
      operationId: find attributes by pet id
      parameters:
        - name: id
          in: path
          description: ID of pet to fetch
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: attribute_id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: pet response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attribute'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /route_not_defined_within_express:
    get:
      description: Returns attributes for this pet
      operationId: route-not-defined-within-express
      parameters:
        - name: name
          in: query
          description: the name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: pet response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attribute'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /pets_charset:
    post:
      description: Creates a new pet in the store.  Duplicates are allowed
      operationId: addPet
      requestBody:
        description: Pet to add to the store
        required: true
        content:
          application/json; charset=utf-8:
            schema:
              $ref: '#/components/schemas/NewPet'
      responses:
        '200':
          description: pet response
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/Pet'
        default:
          description: unexpected error
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
  /pets_content_types:
    post:
      description: Creates a new pet in the store.  Duplicates are allowed
      operationId: addPet
      requestBody:
        description: Pet to add to the store
        required: true
        content:
          application/json; version=1:
            schema:
              $ref: '#/components/schemas/NewPet'
      responses:
        '200':
          description: pet response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        default:
          description: unexpected error
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'

components:
  parameters:
    $ref: 'xt.openapi.parameters.yaml#/parameters'

  schemas:
    $ref: 'xt.newpet.yaml'

    Pet:
      allOf:
        - $ref: 'xt.newpet.yaml#/NewPet'
        - required:
            - id
          properties:
            id:
              type: integer
              format: int64

    NewAttribute:
      required:
        - name
      properties:
        name:
          type: string

    Attribute:
      allOf:
        - $ref: '#/components/schemas/NewAttribute'
        - required:
            - id
          properties:
            id:
              type: integer
              format: int64

    NewPhoto:
      type: object
      required:
        - file
      properties:
        file:
          description: The photo
          type: string
          format: binary

    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string